refactor: move integration specific logic into integrations/[identifier]
