+++ src/lib/hooks/useAllConfigs.ts
@@ -1,8 +1,9 @@
 import { useUser } from "@auth0/nextjs-auth0/client";
 import useSWR from "swr";
 
+import { IntegrationIdentifier } from "@/lib/integrations/active";
 import { fetcher } from "@/lib/utils/fetchUtils";
-import { AllConfigsIndex, IntegrationIdentifier } from "@/types/rezervo";
+import { AllConfigsIndex } from "@/types/rezervo";
 
 export function useAllConfigs(integration: IntegrationIdentifier) {
     const { user } = useUser();
