+++ src/lib/hooks/useUserConfig.ts
@@ -3,8 +3,9 @@ import useSWR from "swr";
 import useSWRMutation from "swr/mutation";
 
 import { useAllConfigs } from "@/lib/hooks/useAllConfigs";
+import { IntegrationIdentifier } from "@/lib/integrations/active";
 import { fetcher } from "@/lib/utils/fetchUtils";
-import { IntegrationConfigPayload, IntegrationIdentifier, IntegrationConfig } from "@/types/rezervo";
+import { IntegrationConfigPayload, IntegrationConfig } from "@/types/rezervo";
 
 function putConfig(url: string, { arg: config }: { arg: IntegrationConfigPayload }) {
     return fetch(url, {
